% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_choice_anycase.R
\name{check_choice_anycase}
\alias{check_choice_anycase}
\alias{assert_choice_anycase}
\title{Check if an object is an element of a given set regardless of case}
\usage{
check_choice_anycase(x, choices, null.ok = FALSE, fmatch = FALSE)

assert_choice_anycase(
  x,
  choices,
  null.ok = FALSE,
  fmatch = FALSE,
  custom.msg = NULL,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{choices}{[\code{character}]\cr
Set of possible values.}

\item{null.ok}{[\code{logical(1)}]\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{fmatch}{[\code{logical(1)}]\cr
Use the set operations implemented in \code{fmatch} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link[base]{match}}.
\code{fastmatch::fmatch} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}

\item{custom.msg}{[\code{character(1)}]\cr
Extra information to be appended to the standard error message in assert-functions}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Depending on the function prefix:\cr
If the check is successful, the function
\code{assert_choice_anycase} return \code{x} invisibly, whereas
\code{check_choice_anycase} return \code{TRUE}.\cr
If the check is not successful,
\code{assert_choice_anycase} throws an error message and
\code{check_choice_anycase} returns a string with the error message.
}
\description{
Check if an object is an element of a given set regardless of the object name is
written in lower and upper case. The function is based on code/{checkmate::check_choice}.
}
\details{
The object must be of type character. The check is intended for functions were using
camelCase may make the argument easier to remember. Therefore, the input to the function is
made case insensitive.
}
\examples{
# returns TRUE
check_choice_anycase(x = "APPLE", 
                     choices = c("Apple", "Pear", "Orange") , 
                     null.ok = FALSE, 
                     fmatch = FALSE)
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
