% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_choice_character.R
\name{check_choice_character}
\alias{check_choice_character}
\alias{assert_choice_character}
\title{Check if an object is an element of a given set}
\usage{
check_choice_character(
  x,
  choices,
  null.ok = FALSE,
  ignore.case = FALSE,
  fmatch = FALSE
)

assert_choice_character(
  x,
  choices,
  null.ok = FALSE,
  ignore.case = FALSE,
  fmatch = FALSE,
  .var.name = checkmate::vname(x),
  comment = NULL,
  add = NULL
)
}
\arguments{
\item{x}{[\code{any}]\cr
Object to check.}

\item{choices}{[\code{character}]\cr
Set of possible values.}

\item{null.ok}{[\code{logical(1)}]\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{ignore.case}{[\code{logical(1)}]\cr
Case is ignored if \code{TRUE}. Default is \code{FALSE}.}

\item{fmatch}{[\code{logical(1)}]\cr
Use the set operations implemented in \code{fmatch} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link[base]{match}}.
\code{fastmatch::fmatch} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{comment}{[\code{character(1)}]\cr
Extra information to be appended to the standard error message in assertions.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Depending on the function prefix:\cr
If the check is successful, the function
\code{assert_choice_character} return \code{x} invisibly, whereas
\code{check_choice_character} return \code{TRUE}.\cr
If the check is not successful,
\code{assert_choice_character} throws an error message and
\code{check_choice_character} returns a string with the error message.
}
\description{
Check if an object is an element of a given set in
the object name. The function is based on \code{checkmate::check_choice},
but includes the argument \code{ignore.case}.
}
\details{
The object must be of type character. The check is intended
for functions were using camelCase may make the argument easier
to remember. Therefore, the function will ignore case if
\code{ignore.case = TRUE}.
}
\examples{
# returns TRUE
check_choice_character(x = "APPLE",
                     choices = c("Apple", "Pear", "Orange"),
                     null.ok = FALSE,
                     ignore.case = TRUE,
                     fmatch = FALSE)
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
