% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_credentials.R
\name{check_credentials}
\alias{check_credentials}
\alias{assert_credentials}
\title{Check if an object is a service for which credentials are saved in the user profile}
\usage{
check_credentials(x)

assert_credentials(
  x,
  custom.msg = NULL,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{custom.msg}{[\code{character(1)}]\cr
Extra information to be appended to the standard error message in assert-functions}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Depending on the function prefix:\cr
If the check is successful, the function
\code{assert_credentials} return \code{x} invisibly, whereas
\code{check_credentials} return \code{TRUE}.\cr
If the check is not successful,
\code{assert_credentials} throws an error message and
\code{check_credentials} returns a string with the error message.
}
\description{
Check if credentials are saved in the user profile for the service.
}
\examples{
check_credentials(x = "PJS")

}
