% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_new_colname.R
\name{check_new_colname}
\alias{check_new_colname}
\alias{assert_new_colname}
\title{Check if an object already is used as a column name}
\usage{
check_new_colname(x, columns, code_column = FALSE, overwrite = FALSE)

assert_new_colname(
  x,
  columns,
  code_column = FALSE,
  overwrite = FALSE,
  comment = NULL,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{columns}{[\code{character}]\cr
The column names of the existing column(s)}

\item{code_column}{[\code{logical(1)}]\cr
Whether the existing column(s) (\code{x}) is the code_column or not.}

\item{overwrite}{[\code{logical(1)}]\cr
Whether existing columns should be overwritten. Only applicable if \code{code_column = FALSE}}

\item{comment}{[\code{character(1)}]\cr
Extra information to be appended to the standard error message in assert-functions}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Depending on the function prefix:\cr
If the check is successful, the function
\code{assert_NewColname} return \code{x} invisibly, whereas
\code{check_NewColname} return \code{TRUE}.\cr
If the check is not successful,
\code{assert_NewColname} throws an error message and
\code{check_NewColname} returns a string with the error message.
}
\description{
Used in the add-functions to check if the name of a new column already exists as column name in
the data frame.
}
\details{
Compares the names of existing column(s) in data frame with the new names of the columns that should be created.
If a new column is the same as the code column, an error is issued. Use the argument \code{code_columne = TRUE} to
indicate that the argument to the columns are the code column.
If the name of the new column already exist but \code{code_column = FALSE}, an error is issued unless \code{overwrite = TRUE}.
}
\examples{
# error:
  check_new_colname(x = "komnr",
                    columns = c("saksnr", "komnr"),
                    code_column = FALSE,
                    overwrite = FALSE)
                    
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
