% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_non_null.R
\name{check_non_null}
\alias{check_non_null}
\alias{assert_non_null}
\title{Check if two or more arguments are NULL or NA}
\usage{
check_non_null(x)

assert_non_null(x, .var.name = checkmate::vname(x), comment = NULL, add = NULL)
}
\arguments{
\item{x}{[any]\cr
List with objects to check.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{comment}{[\code{character(1)}]\cr
Extra information to be appended to the standard error message in assertions.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Depending on the function prefix:\cr
If the check is successful, the function
\code{assert_non_null} return \code{x} invisibly, whereas
\code{check_non_null} return \code{TRUE}.\cr
If the check is not successful,
\code{assert_non_null} throws an error message and
\code{check_non_null} returns a string with the error message.
}
\description{
Check whether two or more arguments that may replace each other both are NULL or NA.
}
\details{
In some functions, one argument may replace another. The check is used to ensure that at least on of the arguments have a value. Specific checks for each argument should be used in addition.
}
\examples{
data <- NULL
nrows_in_data <- 56
check_non_null(x = list(data, nrows_in_data))
}
