% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rodbc.R
\name{check_rodbc}
\alias{check_rodbc}
\alias{assert_rodbc}
\title{Check if an object is an open RODBC-channel}
\usage{
check_rodbc(x, dbservice = NULL)

assert_rodbc(
  x,
  dbservice = NULL,
  custom.msg = NULL,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{dbservice}{[character(1)]\cr
The database the channel should be connect to.}

\item{custom.msg}{[\code{character(1)}]\cr
Extra information to be appended to the standard error message in assert-functions}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Depending on the function prefix:\cr
If the check is successful, the function
\code{assert_rodbc} return \code{x} invisibly, whereas
\code{check_rodbc} return \code{TRUE}.\cr
If the check is not successful,
\code{assert_rodbc} throws an error message and
\code{check_rodbc} returns a string with the error message.
}
\description{
Check if an object is an open RODBC-channel.
}
\details{
The check uses \code{RODBC:::odbcValidChannel} to check if the argument is an open RODBC-channel.
Be aware that this function is an internal \code{RODBC}-function and may change within that package without warning.
}
\examples{
\dontrun{
#login to PJS
journal_rapp <- login("PJS")

# error check:
check_rodbc(x = "journal_rapp", dbservice = "PJS")
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
