% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWarningFunction.R
\name{makeWarning}
\alias{makeWarning}
\alias{makeWarningFunction}
\title{Turn a Check into an Assertion}
\usage{
makeWarning(x, res, var.name)

makeWarningFunction(
  check.fun,
  c.fun = NULL,
  coerce = FALSE,
  env = parent.frame()
)
}
\arguments{
\item{x}{[\code{any}]\cr
Object to check.}

\item{res}{[\code{TRUE} | \code{character(1)}]\cr
The result of a check function: \code{TRUE} for successful checks,
and an error message as string otherwise.}

\item{var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}

\item{check.fun}{[\code{function}]
Function which checks the input. Must return \code{TRUE} on success and a string with the error message otherwise.}

\item{c.fun}{[\code{character(1)}]
If not \code{NULL}, instead of calling the function \code{check.fun}, use \code{.Call} to call a C function \dQuote{c.fun} with the identical
set of parameters. The C function must be registered as a native symbol, see \code{\link[base]{CallExternal}}.
Useful if \code{check.fun} is just a simple wrapper.}

\item{coerce}{[\code{logical(1)}]\cr
If \code{TRUE}, injects some lines of code to convert numeric values to integer after an successful assertion.
    Currently used in \code{\link[checkmate:assertCount]{assert_count}},
    \code{\link[checkmate:assertInt]{assert_int}} and
    \code{\link[checkmate:assertIntegerish]{assert_integerish}}.}

\item{env}{[\code{environment}]
The environment of the created function. Default is the \code{parent.frame}, see \code{\link[base]{sys.parent}}.}

\item{collection}{[\code{\link[checkmate:AssertCollection]{AssertCollection}}]\cr
If an \code{\link[checkmate:AssertCollection]{AssertCollection}} is provided, the error message is stored
in it. If \code{NULL}, an exception is raised if \code{res} is not
\code{TRUE}.}
}
\value{
\code{makeAssertion} invisibly returns the checked object if the check was successful,
 and an exception is raised (or its message stored in the collection) otherwise.
 \code{makeAssertionFunction} returns a \code{function}.
}
\description{
\code{makeAssertionFunction} can be used to automatically create
    an assertion function based on a check function (see example). This is a
    modification of \code{checkmate::makeAssertionFunction} that includes the
    argument comment in the assertion function.
}
\details{
The code is imported from \code{checkmate}. The modifications in the
    code is marked. The argument use.namespace is deleted as
    \code{checkmate::makeAssertion} and \code{checkmate::vname} always should
    be used.
}
\examples{
\dontrun{
# Simple custom check function
checkFalse = function(x) if (!identical(x, FALSE)) "Must be FALSE" else TRUE

# Create the respective assert function
assertFalse = function(x, .var.name = vname(x), add = NULL) {
  res = checkFalse(x)
  makeAssertion(x, res, .var.name, add)
}

# Alternative: Automatically create such a function
assertFalse = makeAssertionFunction(checkFalse)
print(assertFalse)
}

}
