% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIcheckmate-package.R
\docType{package}
\name{NVIcheckmate-package}
\alias{NVIcheckmate}
\alias{NVIcheckmate-package}
\title{NVIcheckmate: Extension of checkmate with argument checking adapted for NVIverse}
\description{
Extends `checkmate` with functions for argument checking that are adapted for NVIverse. NVIcheckmate is intended to be used together with `checkmate`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/NVIcheckmate}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/NVIcheckmate/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{Petter.Hopp@vetinst.no} (\href{https://orcid.org/0000-0002-8695-0378}{ORCID})

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (Creator and copyright holder of the package checkmate from which some code is imported) [copyright holder]
}

Other contributors:
\itemize{
  \item Norwegian Veterinary Institute [copyright holder]
  \item Dénes Tóth \email{toth.denes@kogentum.hu} (Contributor to the assert-function in the package checkmate) [contributor]
}

}
\keyword{internal}
