% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_rows.R
\name{check_duplicate_rows}
\alias{check_duplicate_rows}
\alias{assert_duplicate_rows}
\alias{warn_duplicate_rows}
\title{Check if data has duplicate rows}
\usage{
check_duplicate_rows(x, group_var)

assert_duplicate_rows(
  x,
  group_var,
  .var.name = checkmate::vname(x),
  comment = NULL,
  add = NULL
)

warn_duplicate_rows(
  x,
  group_var,
  .var.name = checkmate::vname(x),
  comment = NULL,
  add = NULL
)
}
\arguments{
\item{x}{[\code{any}]\cr
Object to check.}

\item{group_var}{[\code{character}]\cr
The variables in the data that should constitute an unique id. Defaults to
    \code{NULL}.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{comment}{[\code{character(1)}]\cr
Extra information to be appended to the standard error message in assertions.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Depending on the function prefix:\cr
 If the check is successful, the function 
 \code{assert_duplicate_rows} return \code{x} invisibly, whereas
 \code{check_duplicate_rows} return \code{TRUE}.\cr
 If the check is not successful, 
 \code{assert_duplicate_rows} throws an error message and 
 \code{check_duplicate_rows} returns a string with the error message.
}
\description{
Check if data has duplicate rows for one or more group variables.
}
\details{
Used the check that the group variables can be a unique identifyer
    of rows in the data. The data can be data.frame, tibble or data.table.

The function is mainly intended to be used to check data in scripts
    transforming data. In particular, it is intended for use routine scripts
    preparing data to avoid that there are more than one row for group
    variables that should be unique, for example in data with
    "prodnr8" properties where "prodnr8" is expected to by an unique id. This
    can be important to avoid problems in for example joins/mergings.
}
\examples{
\dontrun{
# Creating data
pt_data <- as.data.frame(list("prodnr8" = c("01234567", "12345678"), "P157" = c(200, 300)))
prod_name <- as.data.frame(
    list("prodnr8" = c("01234567", "12345678", "12345678"),
         "prodnr10" = c("0123456713", "1234567814", "1234567826"),
         "navn" = c("Ola Nordmann", "Kari Trestakk", "Per Trestakk")))
# merge
pt_data <- merge(pt_data, prod_name, by = "prodnr8")
# checking for double rows per prodnr8
assert_duplicate_rows(x = pt_data,
                        group_var = "prodnr8",
                        comment = "Checked after merging pt_data med prod_name")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
